# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: G Creator Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/categories:
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a new category (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryItem"
        "400":
          description: Validation error
        "409":
          description: Already exists
  /admin/categories/{id}:
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
        "404":
          description: Not found
...
  /posts/{postId}/bookmark:
    post:
      operationId: bookmarkPost
      tags: [bookmarks]
      summary: Bookmark a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
    delete:
      operationId: unbookmarkPost
      tags: [bookmarks]
      summary: Remove bookmark
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
...
  /bookmarks:
    get:
      operationId: getBookmarks
      tags: [bookmarks]
      summary: Get bookmarked posts
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

  # PAID CONTENT
...
  /hashtags/trending:
    get:
      operationId: getTrendingHashtags
      tags: [hashtags]
      summary: Get trending hashtags
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hashtag"
  /search:
    get:
      operationId: search
      tags: [hashtags]
      summary: Search posts and users
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [all, posts, users]
            default: all
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"

  # NOTIFICATIONS
...
  /listings:
    get:
      operationId: listListings
      tags: [marketplace]
      summary: List marketplace listings
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sellerId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingList"
    post:
      operationId: createListing
      tags: [marketplace]
      summary: Create a listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
  /listings/{listingId}:
    get:
      operationId: getListing
      tags: [marketplace]
      summary: Get a listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
    patch:
      operationId: updateListing
      tags: [marketplace]
      summary: Update or archive a listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListingInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
  /listings/{listingId}/purchase:
    post:
      operationId: purchaseListing
      tags: [marketplace]
      summary: Purchase a listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders:
    get:
      operationId: listOrders
      tags: [marketplace]
      summary: Get my purchase history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  # AUCTIONS
...
components:
  schemas:
...
    UserSummary:
      type: object
      required: [id, username, displayName]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        accountTier:
          type: string
          enum: [free, creator, pro, elite]
          default: free
        idVerificationStatus:
          type: string
          enum: [none, pending, approved, rejected]
          default: none
        isFollowing:
          type: boolean
        followersCount:
          type: integer
...
    LinkPreview:
      type: object
      required: [title, domain, url]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        domain:
          type: string
        url:
          type: string
    Post:
      type: object
      required: [id, authorId, author, type, caption, likesCount, commentsCount, createdAt]
      properties:
        id:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/UserSummary"
        type:
          type: string
          enum: [text, photo, video, gif, link]
        caption:
          type: string
        mediaUrl:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        muxPlaybackId:
          type: ["string", "null"]
        embedUrl:
          type: ["string", "null"]
        hashtags:
          type: array
          items:
            type: string
        likesCount:
          type: integer
        commentsCount:
          type: integer
        viewsCount:
          type: integer
        isLiked:
          type: boolean
        isBookmarked:
          type: boolean
        isPinned:
          type: boolean
        contentRating:
          type: string
          enum: [safe, suggestive, mature, nsfw, explicit]
          default: safe
        visibility:
          type: string
          enum: [public, followers, subscribers_only]
          default: public
        isLocked:
          type: boolean
          default: false
        price:
          type: ["number", "null"]
        allowDownload:
          type: boolean
          default: false
        downloadPrice:
          type: ["number", "null"]
        isUnlocked:
          type: boolean
          default: false
        displayAspect:
          type: ["string", "null"]
        trimStart:
          type: ["number", "null"]
        trimEnd:
          type: ["number", "null"]
        mediaItems:
          type: array
          items:
            type: string
          description: Ordered list of media URLs for carousel posts (photos/GIFs)
        linkPreview:
          $ref: "#/components/schemas/LinkPreview"
        createdAt:
          type: string
          format: date-time
...
    ListingDetail:
      type: object
      required: [id, sellerId, title, price, type, status, salesCount, createdAt]
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        category:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        type:
          type: string
          enum: [digital, service, physical]
        status:
          type: string
          enum: [active, sold_out, archived]
        salesCount:
          type: integer
        seller:
          $ref: "#/components/schemas/UserSummary"
        createdAt:
          type: string
          format: date-time
    ListingList:
      type: object
      required: [listings, total]
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/ListingDetail"
        total:
          type: integer
    CreateListingInput:
      type: object
      required: [title, price]
      properties:
        title:
          type: string
          minLength: 3
        description:
          type: string
        price:
          type: number
          minimum: 0.01
        category:
          type: string
        imageUrl:
          type: string
        type:
          type: string
          enum: [digital, service]
    UpdateListingInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        status:
          type: string
          enum: [active, sold_out, archived]
    Order:
      type: object
      required: [id, buyerId, listingId, sellerId, price, status, createdAt]
      properties:
        id:
          type: integer
        buyerId:
          type: integer
        listingId:
          type: integer
        sellerId:
          type: integer
        price:
          type: number
        status:
          type: string
        listing:
          $ref: "#/components/schemas/ListingDetail"
        createdAt:
          type: string
          format: date-time

    # HASHTAGS
    Hashtag:
      type: object
      required: [name, postsCount]
      properties:
        name:
          type: string
        postsCount:
          type: integer
        trendingScore:
          type: number
    SearchResults:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserSummary"

    # NOTIFICATIONS
...
    CategoryItem:
      type: object
      required: [id, name, sortOrder, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    CreateCategoryBody:
      type: object
      required: [name]
      properties:
        name:
          type: string
        sortOrder:
          type: integer
...